fix uncaught error when no file was uploaded

+++ src/controllers/uploadController.js
@@ -77,7 +77,8 @@ class UploadController extends PageController {
     }
 
     // delete the file from the uploads folder
-    fs.unlink(req.file.path)
+    if(req.file && req.file.path)
+      fs.unlink(req.file.path)
 
     super.post(req, res, next)
   }
